Fixes for our purposes
