+++ src/Model/Table/OauthAuthCodeScopesTable.php
@@ -2,10 +2,15 @@
 
 namespace OAuthServer\Model\Table;
 
+use Cake\ORM\Association\BelongsTo;
+use Cake\ORM\RulesChecker;
 use Cake\ORM\Table;
 
 /**
- * Auth Code Scope Model
+ * Class AuthCodeScopesTable
+ *
+ * @property BelongsTo|OauthAuthCodeScopesTable $AuthCodes
+ * @property BelongsTo|OauthScopesTable $OauthScopes
  */
 class OauthAuthCodeScopesTable extends Table
 {
@@ -17,6 +22,29 @@ class OauthAuthCodeScopesTable extends Table
         parent::initialize($config);
 
         $this->setTable('oauth_auth_code_scopes');
-        $this->setPrimaryKey('id');
+        $this->setPrimaryKey(['auth_code', 'scope_id']);
+
+        $this->belongsTo('AuthCodes', [
+            'className' => 'OAuthServer.AuthCodes',
+            'foreignKey' => 'auth_code',
+            'propertyName' => 'code',
+        ]);
+        $this->belongsTo('OauthScopes', [
+            'className' => 'OAuthServer.OauthScopes',
+            'foreignKey' => 'scope_id',
+            'propertyName' => 'scopes',
+        ]);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function buildRules(RulesChecker $rules): RulesChecker
+    {
+        $rules->isUnique(['auth_code', 'scope_id']);
+        $rules->addCreate($rules->existsIn('auth_code', 'AuthCodes'));
+        $rules->addCreate($rules->existsIn('scope_id', 'OauthScopes'));
+
+        return $rules;
     }
 }
